+++ disneyAPI/src/main/java/com/alkemy/disneyAPI/repositories/GenderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GenderRepository extends CrudRepository<Gender, Integer> {
+public interface GenderRepository extends CrudRepository<Gender, Long> {
 
 }
